+++ server/src/resources/food_menu/food_menu.service.ts
@@ -1,10 +1,15 @@
 import FoodMenuModel from '@/resources/food_menu/food_menu.model';
+import MenuModel from '@/resources/menu/menu.model';
+import AccountFoodModel from '@/resources/account_food/account_food.model';
 import FoodMenu from '@/resources/food_menu/food_menu.interface';
 import { Schema } from 'mongoose';
 import Props from '@/utils/types/props.type';
+import { Menu } from '@/resources/menu/menu.interface';
 
 class FoodMenuService {
     private food_menu = FoodMenuModel;
+    private menu = MenuModel;
+    private account_food = AccountFoodModel;
 
     /**
      * Create a new food menu
@@ -14,9 +19,31 @@ class FoodMenuService {
         food_id: Schema.Types.ObjectId,
         menu_id: Schema.Types.ObjectId,
         count: number,
-        equiv_type: string
+        equiv_type: string,
+        account_id: Schema.Types.ObjectId
     ): Promise<FoodMenu | Error> {
         try {
+            const food_menu_temp = await this.menu.findById(menu_id);
+            const account_food_temp = await this.account_food.findById(food_id);
+
+            if (!food_menu_temp) {
+                throw new Error('Unable to find menu food');
+            }
+
+            if (
+                account_food_temp &&
+                account_food_temp.account_id.toString() !==
+                    account_id.toString()
+            ) {
+                throw new Error('You are not allowed to get this menu food');
+            }
+
+            if (
+                food_menu_temp.account_id.toString() !== account_id.toString()
+            ) {
+                throw new Error('You are not allowed to get this menu food');
+            }
+
             const food_menu = await this.food_menu.create({
                 foodModel,
                 food_id,
@@ -40,9 +67,33 @@ class FoodMenuService {
         food_id: Schema.Types.ObjectId,
         menu_id: Schema.Types.ObjectId,
         count: number,
-        equiv_type: string
+        equiv_type: string,
+        account_id: Schema.Types.ObjectId
     ): Promise<FoodMenu | Error> {
         try {
+            const food_menu_temp = await this.food_menu
+                .findById(_id)
+                .populate({
+                    path: 'food_id',
+                    populate: { path: '_id' },
+                })
+                .populate({
+                    path: 'menu_id',
+                    populate: { path: '_id' },
+                });
+
+            if (!food_menu_temp) {
+                throw new Error('Unable to find menu food');
+            }
+            const menu = food_menu_temp.menu_id as Menu;
+            if (
+                food_menu_temp &&
+                food_menu_temp.menu_id &&
+                menu.account_id.toString() !== account_id.toString()
+            ) {
+                throw new Error('You are not allowed to update this menu food');
+            }
+
             const food_menu = await this.food_menu
                 .findByIdAndUpdate(
                     _id,
@@ -65,12 +116,12 @@ class FoodMenuService {
                 });
 
             if (!food_menu) {
-                throw new Error('Unable to update food menu with thad id');
+                throw new Error('Unable to update food menu with that data');
             }
 
             return food_menu;
-        } catch (error) {
-            throw new Error('Unable to change food menu');
+        } catch (error: any) {
+            throw new Error(error.message);
         }
     }
 
@@ -78,8 +129,34 @@ class FoodMenuService {
      * Attempt to delete food menu
      */
 
-    public async delete(_id: Schema.Types.ObjectId): Promise<FoodMenu | Error> {
+    public async delete(
+        _id: Schema.Types.ObjectId,
+        account_id: Schema.Types.ObjectId
+    ): Promise<FoodMenu | Error> {
         try {
+            const food_menu_temp = await this.food_menu
+                .findById(_id)
+                .populate({
+                    path: 'food_id',
+                    populate: { path: '_id' },
+                })
+                .populate({
+                    path: 'menu_id',
+                    populate: { path: '_id' },
+                });
+
+            if (!food_menu_temp) {
+                throw new Error('Unable to delete menu food');
+            }
+            const menu = food_menu_temp.menu_id as Menu;
+            if (
+                food_menu_temp &&
+                food_menu_temp.menu_id &&
+                menu.account_id.toString() !== account_id.toString()
+            ) {
+                throw new Error('You are not allowed to delete this menu food');
+            }
+
             const food_menu = await this.food_menu
                 .findByIdAndDelete(_id)
                 .populate({
@@ -92,12 +169,12 @@ class FoodMenuService {
                 });
 
             if (!food_menu) {
-                throw new Error('Unable to delete food menu with that id');
+                throw new Error('Unable to delete food menu with that data');
             }
 
             return food_menu;
-        } catch (error) {
-            throw new Error('Unable to delete food menu');
+        } catch (error: any) {
+            throw new Error(error.message);
         }
     }
 
@@ -105,7 +182,8 @@ class FoodMenuService {
      * Attempt to find all food in  menu
      */
     public async get(
-        menu_id: Schema.Types.ObjectId
+        menu_id: Schema.Types.ObjectId,
+        account_id: Schema.Types.ObjectId
     ): Promise<FoodMenu | Array<FoodMenu> | Error> {
         try {
             const food_menu = await this.food_menu
@@ -125,21 +203,39 @@ class FoodMenuService {
                 throw new Error('Unable to find food menu');
             }
 
+            const menu = food_menu[0].menu_id as Menu;
+
+            if (account_id.toString() !== menu.account_id.toString()) {
+                throw new Error('You are not allowed to get this menu food');
+            }
+
             return food_menu;
-        } catch (error) {
-            throw new Error('Unable to find food menu');
+        } catch (error: any) {
+            throw new Error(error.message);
         }
     }
-
 
     /**
      * Attempt to find food menu
      */
 
     public async find(
-        props: Props
+        props: Props,
+        account_id: Schema.Types.ObjectId
     ): Promise<FoodMenu | Array<FoodMenu> | Error> {
         try {
+            const food_menu_temp = await this.menu
+                .find({ account_id: account_id })
+                .select(['_id'])
+                .exec();
+
+            const propertyValues = [] as Array<string>;
+            food_menu_temp.map((item) => {
+                propertyValues.push(item._id.toString());
+            });
+
+            props.menu_id = { $in: propertyValues };
+
             const food_menu = await this.food_menu
                 .find(props, null, {
                     sort: { name: 1 },
@@ -158,8 +254,47 @@ class FoodMenuService {
             }
 
             return food_menu;
-        } catch (error) {
-            throw new Error('Unable to find food menu');
+        } catch (error: any) {
+            throw new Error(error.message);
+        }
+    }
+
+    /**
+     * Attempt to find all account food
+     */
+    public async adminGet(): Promise<FoodMenu | Array<FoodMenu> | Error> {
+        try {
+            const food_menu = await this.food_menu.find({}, null, {
+                sort: { name: 1 },
+            });
+
+            if (!food_menu) {
+                throw new Error('Unable to find food_menu');
+            }
+
+            return food_menu;
+        } catch (error: any) {
+            throw new Error(error.message);
+        }
+    }
+
+    /**
+     * Attempt to delete account food
+     */
+
+    public async adminDelete(
+        _id: Schema.Types.ObjectId
+    ): Promise<FoodMenu | Error> {
+        try {
+            const food_menu = await this.food_menu.findByIdAndDelete(_id);
+
+            if (!food_menu) {
+                throw new Error('Unable to delete food_menu with that data');
+            }
+
+            return food_menu;
+        } catch (error: any) {
+            throw new Error(error.message);
         }
     }
 }
